requirements.txt
pandas
matplotlib
seaborn
python-dotenv
openai
